"7:50am. I am back and so has my mind returned to battle readiness. I've been thinking about my idea of propagating lazy values and decided against it after all.
I am going to go with intermediate nodes, but rather than what I talked about yesterday (ref count 2 nodes,) what I will do are attention modulated switches. One thing I am going to have to figure out is how to get `DistinctByKey` to use reference equality.

8am. Ah, let me try just that particular thing now.

```fsharp
Observable.single 1
|> Observable.distinctKeyCompare id
        {new Collections.Generic.IEqualityComparer<_> with
            member _.Equals(a,b) = a = b
            member _.GetHashCode(x) = hash x
            }
```

Yeah, as expected there is a way to get that functionality. It is no problem at all.

```
Collections.HashIdentity.Reference
```

Ah, I never realized it before, but this is literally an `IEqualityComparer`.

```
        Observable.single 1
        |> Observable.distinctKeyCompare id Collections.HashIdentity.Structural
```

So I can just swap this for the reference comparer wherever I need it. Great.

8:10am. This is great. Just last night I've been thinking how great it would be if I could pass `HashIdentity.Reference` into `Distinct` and as it turns out that is possible.

This is what I meant by reactive combinators being a neverending parade of joy.

8:15am. Now let me talk about those attention modulated switches.

As one could guess I am going to have the editor send me info on which file the user is editing. That is the attention part.

I meant to use lazy values to get efficiency, but what occurred to me is that lazy values would make it really difficult to abort the computation as it is going on. I want typechecking to switch on new info and file changes.

For me to have missed this while I was awake yesterday, I really must have been out of it. I even planned this, but forgot about it in the interim.

Reactive combinators are so great - there are just so many opportunities for abstraction and good design, and proving one's skill.

8:20am. The next thing on my list is something I should have done earlier. Let me open an issue for the spell checker.

8:40am. Er, I opened an PR for the reactive typo. There has been some activity there. I've committed myself to fixing the `switchMap` and `choose` after I finish the GUI examples.

Now the spellchecker.

8:50am. https://github.com/bartosz-antosik/vscode-spellright/issues/
Issue spell checking a "large" document

Somebody already did it. Issue 351.

Ok.

8:55am. Let me take a break here. Machimaho and that Looking Up To Magical Girls TL thread are waiting for me. Last night I was too tired to read them and went to bed at 8pm.

Let me slack here for a while, and then I will go into it full force. If possible it would be good if I could finish the ZeroMQ examples by the end of the month."