"7:30am. I am up this early despite going to bed at 11pm yesterday. Am I looking forward to every programming day or what? May has been great, if exhausting. Today is the last day of the month so it is time to do that monthly review for the PL thread. I'll take some time to do an extended rest today.
And I'll start with my usual morning chilling. Kaiji is out.

https://mangadex.org/title/40770/shinigami-ni-sodaterareta-shoujo-wa-shikkoku-no-tsurugi-wo-mune-ni-idaku

This came out yesterday too. Let me have my fun.

9am. Ok, let me start this thing. Review time.

///

Towards the end of April I had a change of mindset - rather than trying to figure out those massively bloated LSP related projects, I should directly study the VS Code editor itself and aim to directly integrate my language with it. That went well. I spent nearly two weeks studying the various samples and [refactoring them](https://github.com/mrakgr/vscode-extension-samples#vs-code-extensions-refactor) as I went along. And this time I have it. The feedback from this experience was good. This is a sharp contrast to...for example February when I picked up an ASP.NET book and realized that no matter how much I wracked my brain that I did not understand what most of the code in the Tic Tac Toe example did. It was so deluged with crap, that for most of the lines I could only guess at what their purpose was.

Not so here. When things are simple, then not only are they easy to learn, but it is also easy to infer what the purpose of various parts should be. This is important for being able to tell when a bug is occurring in the library.

After the first week, I was done with that and I decided to start work on editor support for Spiral. As the first step, I wrote a parser for config files for it and then looked into integrating it with VS Code. I mean, at that point on paper I had everything set - I am familiar with the VS Code editor and its API on one side, and the way to design the language on the other is clear to me.

When the time came to actually transfer data from one side to the other, I realized I had a problem. If you Google something like [`node ipc`](https://www.npmjs.com/package/node-ipc) you'll stumble upon the namesake package. It allows sending messages via IPC as one would expect. On the .NET side however, you have what? Network streams? What am I going to do with those? Surely you want to at least send messages.

I even opened a thread on the F# sub asking how to send strings across different processes, but the answer was a [dreadful resounding silence](https://www.reddit.com/r/fsharp/comments/gek8cv/what_should_i_be_using_to_do_ipc_for_a_language/).

In programming, it is easy to fall into the trap of thinking that because something is simple that it is also easy.

---

The face of my next challenge having been made clear, after searching for a while I stumbled upon [ZeroMQ](http://zguide.zeromq.org/) and decided that I need to master it. And what better way to go about it than to weave all the strands I've been picking up over past few months together. I wanted to break my dependency on the command line for a long time, and given my recent successes with [reactive combinators](http://reactivex.io/) this seemed like an idea time to do some serious UI work and finally master that aspect of programming as well. I failed at this in 2016, but at this point I am too overleveled for that to happen. As a programmer, I've always had this insecurity - I've always had the raw talent, and in the past 5 years the consistent effort I've putting in has refined it into a gem, but there are still some impurities left to polish. My ability to produce good tools is still too poor for me to call myself a real expert. It is one thing to be able to produce good, efficient and readable code in some places, but it quite another to have such a broad base that one can excel regardless of the domain.

The way to break into such a level is the question, and the answer is concurrency.

The only bright spot in the past five months was when I decided to really commit to studying reactive combinators and did some [simple UIs](https://github.com/mrakgr/Lithe-POC) using them. May was just so great. It was fantastic.

It was like learning programming again - every session that I had was meaningful, and the things I've learned will not be forgotten and will serve as the core of my skill from here on out. Reactive combinators are just divine, they are as good for concurrency as parser combinators are for parsing. Master them, add agents to the mix (and [maybe channels](https://github.com/Hopac/Hopac/blob/master/Docs/Programming.md)) and things like UIs suddenly become tractable. And webdev which has been a big thing for a while now is fundamentally very similar to UI work. Though I do not plan on doing webdev, this is what one needs to know if one wants to do it. HTML, CSS, JavaScript, TypeScript and big frameworks on whatever platform are just distracting trivialities in the challenge of producing good abstractions for user experience. Reactive combinators are what is fundamental. Concurrency is what should be the true object of study. The impression that concurrency is about programming with [impossible to reason about](http://deadlockempire.github.io/) low level primitives needs to die.

The reason that I decided to start with frameworks in February was because I lacked the skills and the confidence to manage complexity in those domains.

I finally see the correct path now - all I need to do is keep following it. In a week or so I will be done [translating ZeroMQ examples](https://github.com/mrakgr/Lithe-POC/blob/9bd84b3b18fa0c0cc29175a1d17ba725c1b02a83/Lithe%20Avalonia%20Examples/05_ZeroMQ.fs) from the guide into NetMQ, and then I will finally be able to start work on editor support for Spiral. This time there won't be any more challenges to get in my way. As a rough guess I'd say it will take me 3 months to make v0.2 + editor support for it. Considering how much v0.09 took me and how much better v0.2 is going to be than it, I consider this a bargain.

After that it will be time to get some [new toys](https://www.youtube.com/watch?v=jhQgElvtb1s). I will resume the hunt for the AI that I left off in 2018. Sure, I might not have particularly excelled at this so far, but the cool thing about this quest is that even if you fail, as long you do not give up every 5 years you will get the chance to try again with 100x more computing power than the last time. Personally, I'll aim to top that off with broader programming skills and better tools as well.

This strategy is bound to work for me at some point.

///

10:20am. I am checking out the ZeroMQ guide as I write the above, but it seems that it is down. If this continues, I'll try to grab something off Libgen.

11:10am. I am done with the review.

11:30am. Doing some proofreading and editing. I think what I have now should be decent. It is a bit long, but it is fine if I break into story at some points.

11:40am. I've been thinking about talking about NetMQ, but who cares about those trivial details. For some people, the stories are about the characters, but for me they are about transmiting the emotional valence. They are about the world itself and power.

The above review is good. I crap on some things and praise others, all deservingly. Imagine if I had ready this a few months ago - it would have saved me quite a bit of time.

Let me commit here and then I'll have breakfast and do some chilling. After that it is review time again."