"8:40am. I am up. Let me chill for a bit and then I will start.
9am. Just a bit more and I'll start. It turns out am a few chapters behind on GS and its spin off, but I'll leave that for later.

9:20am. Let me start. It is time to get this going again. I want to do a few examples today. It would really be great if I could do one example per hour, but I keep getting stuck on strange things that take me a long time to figure out.

9:25am. 173/515. I'll finally figure out what heartbeating is.

9:55am. I've been reading the example this one will be tough. I'll have to take a look at the F# example as well as C code is tedious.

10am. I do not really understand what to do if the balancer is down. What should the worker do then?

10:35am. I am thinking about this.

https://github.com/booksbyus/zguide/blob/master/examples/C%23/ppqueue.cs

It does not seem like there is a F# version of this, but both in the book and here, the code is such a mess. At this point even more than merely implementing the thing, I am wondering what a good example of this should be like.

10:55am. Had to take a little break.

11:05am. I am still thinking about it.

11:10am. And thinking about it...

I have a problem with heartbeating. The main issue is that the existing mechanism of how this works is already a good way of doing it. If the worker aborts, then the router will just stop sending to it.

Just how is the router supposed to restart the worker? I have no idea.

I hadn't architected for this kind of thing at all.

11:25am. I am doing some really hard thinking here.

I am starting to hit the limits of what can be done with ZeroMQ. I can't really do systems much more complex than this. Because everything is untyped, I am constantly writing my own serialization. That is in essence all I've been doing for the past few weeks. I can't really do this for much more than the toy example here.

In Spiral v0.2 when it comes around, I would be able to take a library like NetMQ and make it much more useful by having it autogenerate generic receives for arbitrary union types.

There is something I am wondering about.

Thinking back to that routing example was it really the case that it did not have zero delimiters for multi step messages?

```
let show (msg : NetMQMessage) = Seq.toArray msg |> Array.map (fun x -> x.ConvertToString()) |> String.concat " | "
```

Let me change this a little so as to highlight the nulls.

```
let show (msg : NetMQMessage) = Seq.toArray msg |> Array.map (fun x -> if x.IsEmpty then "<null>" else x.ConvertToString()) |> String.concat " | "
```

11:35am. Yeah, there are indeed empty frames between the addresses.

The way the ZeroMQ message format is structured makes separating the data from the addresses really challenging.

You really cannot use multi part messages for data if you want to do any routing due to the empty strings being the same as frame delimiters.

Instead, I would need my own parsers for parsing submessages.

11:40am. It is the C legacy of it. In a higher level language ZeroMQ could be much nicer.

...Let me check out F#'s binary serialization capabilities.

https://mbraceproject.github.io/FsPickler/

Oh this is interesting.

https://mbraceproject.github.io/FsPickler/tutorial.html

Wow, this is actually pretty fantastic. How did it occur to me to look for this only now?

12pm. Change of plans. I am going to stop the NetMQ examples for a bit so I can play with the F# pickler. This is one of those super essential technologies that I need to master much like ZeroMQ itself and reactive combinators and agents.

I could make ZeroMQ much more useful with better binary serialization. Was it a few days ago that I wracked my head how to get around the limitations impossed to me by the way ZeroMQ did addressing?

The solution for all those difficulties is good serialization.

I've thought about whether it would be possible to do better than what I did in v0.09 of Spiral. It should be, but I've been drawing a blank until now on how to get more out of it. I think that studying this library should give me some inspiration. It is really a mystery to as to how to do pickling with top-down type system language like F#.

I am really glad other people put thought into this.

12:10pm. I am going to stop focusing so much on how exactly to do the paranoid pirate, and use the example as an opportunity to showcase some extra features of the NetMQ library.

On crashes, instead of aborting the thread, I am simply going to have it go into a state where it throws away everything except a restart message.

Also I am going to go back to the two queue way of doing things. What I want to do is filter out the requests from the clients so that only the latest are processed. I'll a stack and filter out the old requests.

Also rather than reconnecting the request socket on the every retry, I am going to have the clients have more stable identities and use a single dealer socket for everything per client. Constantly making new sockets creates new identities for them, so when there is overload this creates a lot of contention.

12:15pm. Let me stop here. As hard as it is to believe, I did not put down a line of code in the morning.

Going through the examples is not trivial and I am not doing a straightforward translation, but instead exploring the full capabilities of the library along the way. Like now, I've had difficulties scaling the library due to its untypedness, and have all of a sudden discovered pickling.

No doubt I am going to be looking under the hood for the pickling library as well. I am going to have to do that in order to figure out how to generate the Typescript code for deserialization on the other side. This will take some time.

But aiming for nothing less than full mastery."