"2:25pm. Done with breakfast. Let me slack a bit more.
2:55pm. Ok, enough. I need to start this thing.

...Hmmm, I do want to read up a bit more on MPL. So let me do that.

3pm. https://www.youtube.com/watch?v=cJIi-hIlCQM
CppCon 2015: Kevin P. Fleming “A Crash Course in Open Source Licensing"

This 1.5 hours. If I watch this I won't have much left in me for programming, but maybe I should in fact just go with that.

https://speckyboy.com/which-license-do-you-choose-for-your-code/
https://tldrlegal.com/license/mozilla-public-license-2.0-(mpl-2)

Let me do it.

Let me get this out of my system.

> MPL is a copyleft license that is easy to comply with. You must make the source code for any of your changes available under MPL, but you can combine the MPL software with proprietary code, as long as you keep the MPL code in separate files. Version 2.0 is, by default, compatible with LGPL and GPL version 2 or greater. You can distribute binaries under a proprietary license, as long as you make the source available under MPL.

What does compatibility with LGPL mean here?

> This license is mainly applied to libraries. You may copy, distribute and modify the software provided that modifications are described and licensed for free under LGPL. Derivatives works (including modifications or anything statically linked to the library) can only be redistributed under LGPL, but applications that use the library don't have to be.

The descriptions on this site are better.

3:10pm. https://en.wikipedia.org/wiki/License_compatibility

Never really understood what this meant specifically.

3:15pm. Let me watch the video. The article is poorly written and I am having trouble focusing on it.

...Actually, I forgot that I need to do chores. I'll be right back.

4:10pm. https://youtu.be/cJIi-hIlCQM?t=2344

The talk is informative, even though I am falling asleep by now. Here he talks about the requirement to document changes. The GPL licenses require that and I've been wondering how that would go before the era of Git since it would be so onerous to actually do.

He mentions that this is the least enforced clause in all of open source. That answers my question. Probably people were ignoring that requirement. So it is fine that MPL does not have it.

https://youtu.be/cJIi-hIlCQM?t=2524

This part covers static linking.

5pm. https://youtu.be/cJIi-hIlCQM?t=4656
> If you make certain types of changes you are allowed to make to the software and still call it Firefox, but if you go outside the scope of those changes you are not allowed to call it Firefox anymore. That's why in Debian it is not called Firefox.

5:35pm. Ok, enough of this. I've spend the whole day immersed in license issues. This is as good as I am going to get without going to law school.

I said I would wait two days, but let me change the license of Spiral to MPLv2 now. If /u/frumious or anybody else chimes in that thread, I'll be open to changing it. It is not like Spiral has any users by myself at the moment, so this is really easy to do. No problem at all.

5:40pm. From here on out, I'll try not to think about this anymore. Tomorrow, I want to focus on making progress on editor support for Spiral. Tomorrow, I will try to integrate the config parser with VS Code. I have everything I need for that, though I do not have quite everything ready for full support on the Node side yet. I wish that I get replies to those ZeroMQ.js issues soon so I can figure out how to set up router and dealer sockets and do routing on them.

Well, that will have to wait till tomorrow.

Right now, into the clipboard the MPL goes. And done.

5:50pm. So to reiterate, it was a tossup between LGPL and MPL. I choose the later due to its more permissive compilation restrictions. LGPL would force anybody using Spiral as a library to link it dynamically and I do not want to block any compilers from doing whole program optimization much like Spiral does. Picking LGPL would give me issues down the road if Spiral was rewritten in itself - because Spiral links everything statically, LGPL makes no sense for either it or its libraries.

I am really grateful to that SO question for making making these library linking differences between the two licenses clear to me. I am not sure whether I would have inferred it on my own. It would have taken more effort to be sure.

6pm. Interesting that the CppCon speaker said that there is no definite yes when it comes to law. So I am doing exactly the right thing by seeking a wide array of opinions here.

https://www.reddit.com/r/opensource/comments/5kjwlp/thinking_of_switching_to_mozilla_public_license/

Let me read some more. I'll call it a day soon.

6:10pm. Not much there on Reddit. I am not surprised, it is not exactly a thrilling subject.

6:30pm. I am reading the license itself to see if it is consistent with what I know. I would not be able to figure out that distributing the source when doing compilation is not necessary just by reading this.

> However, a larger work using the licensed work may be distributed under different terms and without source code for files added in the larger work.

Actually, reading this more closely, doesn't this mean that the original MPL library has to have its source disclosed even if it is used as part of proprietary work?

6:50pm. I'll interpret it like that.

> You may create and distribute a Larger Work under terms of Your choice, provided that You also comply with the requirements of this License for the Covered Software.

This is from the license. The said requirements being that the source of the Covered Software is disclosed. The same applies for Covered Software being distributed in Executable Form.

6:55pm. Let me affix this view in my mind.

Yes, I know that I should have had more patience and did this before setting the license to MIT, but the way I roll is to try out a move and then dwell on it afterward. It seems to be intrinsic to my personality. I do this to save my mental bandwidth.

When I set the license my mental model was the larger works could be completely opaque, but that makes no sense. Even though they might be statically linked programs that use MPL libraries, they are still required to disclose the source of the MPL libraries they are using.

This is something I will also need to keep in mind when using libraries in commercial software.

I need to keep in mind that copyleft ~= disclose source. If they chain got broken due to compilation, then there would not be much point to the license now wouldn't it? It would be no different from Apache.

https://softwareengineering.stackexchange.com/questions/174907/can-i-use-epl-licensed-libraries-and-not-give-out-the-source-code-of-my-applicat

Eclipse was one of the licenses suggested on the FSF page. Even though it is a weak-copyleft license it does require disclosing of the source.

7pm. Let me force myself to stop here. I meant to do it 1.5 hours ago. As usual once I get going, I find it hard to stop.

Tomorrow, it is language server time.

I will focus my whole being on getting that config parser done. If I do not get a reply from the authors on how to do use router sockets, nor polling, I am going to have to spend time on that myself. Strictly speaking I would not need to, but I can't feel at ease unless I do things properly and in turn.

I have so much work left to do before I master all of this, but it should be doable by the end of the year. By the end of the year I will have it all. Since I know that, I feel a strong urge to keep going.

I will keep going until it is done.

No matter what, I do not want to let go of this feeling. Some losers might cry and moan about missing the best parts of history, but to me the best part is what is coming up next. This decade will be a blast."