"9:20am. I could have gotten up earlier, but there was some bad weather that I needed to wait for it to clear.
9:25am. Let me do my reading and then I will start.

And holy shit, nobody is answering my issues. What a load of crock. I might need to dive into the code myself at this rate.

10:10am. Let me start.

Today I somehow did not have enough time in the morning to enjoy myself properly, but that does not matter.

10:15am. Let me do some work. There will be plenty of time to lead the Overlord epilogue during the break.

10:20am.

```fs
type ClientMsgs =
    | ProjectFile of {|spiprojDir : string; spiprojText : string|}
    | File of {|spiprojDir : string; text : string|}
```

The first step is here.

10:25am. I need some time to think about it.

10:30am. Hmmm...forget the stuff on the VS Code side for a while. Already I am confronted with the full difficulty of doing that reactive pipeline.

```fs
        | ProjectFile x ->
            msg.Push(config x.spiprojDir x.spiprojText |> process_errors |> Json.serialize)
```

I won't worry about the exact way I want to pass the schema. Instead I need to do the result right now.

10:35am. This is going to be tough. Reactive combinators sort of give me what I need, but not completely. It is hard. I think that I might need to write my own combinators. And that means fully mastering them. Right now I am only at the intermediate stage. Doing the reactive pipeline won't be easy, but it will be something to separate the me of now from who I was before.

Spiral v0.2 is my message to the world that I am moving to the next stage.

10:40am. Now let me mop up some side stuff. I need to also ask how to dispose Req sockets.

10:45am. Writing that thing out is making me realize that the equivalent NetMQ version of the client would just block forever. So the way to do disposal would be to add timeouts.

```
const client = async (spiprojDir: string, spiprojText: string) => {
    const sock = new zmq.Request();
    sock.receiveTimeout = 30e3;
```

That should take care of that.

11am. I am thinking. I am thinking.

I am completely serious here.

What I need to do the reactive pipeline are reactive combinators, but they need to be completely hot everywhere.

I need to generalize that thing I did in the UI with the agent. I essentially want agents that are subscribable. And I do not need `onComplete` and `onError`. I also do not need every agent to have its own thread.

Actually, it is not that I do not need `onError`. It is more like I do not want it to mute the pipeline.

11:10am. Let me take a little break. What I should be doing here is prototyping.

After that I should study the Rx source again. I really should get to the bottom of each of the questions that I had for it.

11:40am. I am sitting in this chair just thinking. I need more - more skill. I know all the features that the pipeline should posses. That has been worked out.

Likewise, my technique and conceptual understanding when it comes to push concurrency is solid. But I am not all the way there yet. I need more knowledge.

11:45am. I need to think about it more. Right now I've yet to even form the basic goals on how to overcome this block.

It is fine. None of this is trivial. I am not going to rush this. I should take my time to attain complete and absolute mastery of push based concurrency.

Pretty much everything I want with this pipeline falls between the cracks that existing techniques do not cover.

While doing the UI I've really been hurt quite a bit by the library not having `mapFold` and better support for hot observables.

When it comes to doing hot observables, the first thing that comes to mind is `distinct` and its variants. I should definitely study that.

12pm. It is already this late. I am in a hot state where I cannot move forward, and I cannot move back. Doing this reactive pipeline will be another great challenge for me.

Yeah, 2020 is the year of concurrnecy for me, much like 2019 was the year of mathematics and 2018 the year of staged functional programming. Every year has its theme. I should take my time to savor it.

12:05pm. Let me force the off switch for a bit here. There will be plenty of time to wrack my brain once it is time to resume.

Today it does not look like I will be doing any programming. Maybe I will step away from the screen to just think. I should take this challenge slowly."