"9:20am. I am up.
9:40am. Let me just go for Ribbon Warrior and then I'll start. Honestly, I actually feel like reading the C# book more. Yesterday I read a few pages forward, and it started talking about user and kernel mode constructs which feels like something I should know about.

9:55am. That was fun. Let me go for the C# book. Then I'll work on the reactive pipeline. Baby steps and all that. I'll get through the block as long as I keep putting in effort. So that is what I am going to do.

10am. I am not going to stress myself over the progress I am making or set deadlines. I'll just do good 5h every day.

> The System.Threading namespace offers an abstract base class called WaitHandle. The WaitHandle class is a simple class whose sole purpose is to wrap a Windows kernel object handle.

I had no idea about this.

10:05am.

> A common usage of the kernel-mode constructs is to create the kind of application that allows only one instance of itself to execute at any given time.

Hmmmm, actually this could be useful for Spiral. Instead of starting the process explicitly, having only one would be good.

```cs
public static class Program {
    public static void Main() {
        Boolean createdNew;

        // Try to create a kernel object with the specified name
        using (new Semaphore(0, 1, "SomeUniqueStringIdentifyingMyApp", out createdNew)) {
            if (createdNew) {
            // This thread created the kernel object so no other instance of this
            // application must be running. Run the rest of the application here...
            } else {
            // This thread opened an existing kernel object with the same string name;
            // another instance of this application must be running now.
            // There is nothing to do in here, let's just return from Main to terminate
            // this second instance of the application.
            }
        }
    }
}
```

Interesting. Let me take a look at this. Yeah, I see now that the non-slim version of these constructs have names, and the `createdNew` parameters.

This is really great. Already reading the book is paying off.

10:30am. 823/896. So it turns out there is a reason why all the locking code allocates that extra object rather than using the class itself. I've been wondering about this!

10:40am.

> It should be clear from this discussion that Monitor should not have been implemented as a static class; it should have been implemented like all the other locks: a class you instantiate and call instance methods on.  In fact, Monitor has many other problems associated with it that are all because it is a static class.

Interesting.

>  The recommendation is not to use C#’s lock statement.

Oh, lol.

Holy shit, is the Monitor broken. They added an extra field for every object into the CLR, but it turned out to be a huge mistake.

11am.

> The System.Threading.Barrier construct is designed to solve a very rare problem, so it is unlikely that you will have a use for it. Barrier is used to control a set of threads that are working together in parallel so that they can step through phases of the algorithm together.

Reminds me of Cuda's `threadsync`. In fact it is probably exactly like that.

11:20am. 834/896. Wow, this stuff is tricky, but I do get it.

Yeah, it is fine if I just finish this book for today.

The annoying thing is that all this complexity ends up being introduced due to compiler optimizations. It is ridiculous.

11:55am. 847/896. The rest is the appendix.

With this I am done with it.

I feel that my understanding of concurrency primitives went up a notch thanks to all of this. Today's morning and yesterday were much better than I expected they would be. I thought that concurrency would be a grindfest, but I am getting useful understanding out of this.

12:05pm. Now, let me stop here for a while. When it is time to resume, I will definitely do some prototyping.

No more scouting. The next move will be to start the attack."