"6:50pm. Done with lunch. There has been thunder, so it is good that I stepped away for a while anyway.
I've really thought deeply into it. And as a result, my thinking is moving away from trying to express the problem through reactive combinators, and more through finite state machines that I could implement using agents.

I realized something - suppose I am doing typechecking - I do not want to be doing typechecking on more than one file at a time. This is for efficiency reasons. But this would very much turn out to be the case if I were expressing the model through reactive combinators.

Instead of doing that, it is better to have a single agent managing at most one task at a time for the whole chain of typechecking. The same will go for compilation.

The only exception will be parsing - that would do well with multiple concurrent tasks as parsing does not depend on previously done work.

I need to be able to manage work explicitly - if typechecking is being done on file 5 and the user makes a change on file 2 then I do not want to wait till it finished. It should immediately drop what it is doing and start work on file 2.

If the user makes a change of file 6 or later, then it is fine, the typechecking can go on without interruption.

7pm. It would be hard to do this if I am dealing with observable chains.

In a similar fashion, it would be hard to deal with nested chains of observables, but this would be easy with agents.

In the first place, observables make sense when you have a transformation of a -> b, but here I am just sending messages rather doing linear transformations. I need to make it efficient and that is a poor fit for the reactive combinators. What I am going to do will still be push concurrency, but it will be of a different sort.

7:20pm. Let me stop here for the day. I have the design that I need in mind. I should be able to implement it tomorrow.

This should go well.

It is kind of a pity that all the thinking that I've put into it before was not enough to converge me to the correct answer right away, but even so the good thing about high level thinking is that you can reuse the elements of it. I haven't wasted this time.

Tomorrow I should be able to do something substantial."