"8:15am. During the night I've been laser focused on what I want to do. But before I start today let me chill for a while.
8:35am. Ok, I am ready...to start reading those Her Majesty Swarm volumes I found yesterday night. Let me do that for a while - and then I will start programming.

9am. Uf, I haven't even started reading it. Just 10m and then I'll start programming.

9:30am. Ok, let me start. I indulged myself a bit. Let me see if I can do the prototype before the morning is over.

Every person (of intelligence) in this world contributes something to it at least once. And my contribution to it will definitely be staged functional programming in the form of Spiral. My design sense is the thing I need to share.

I'll do this only once. As a human I do not feel like doing more than this. I am supposed to be selfish and lazy, but here I am spending my days working for free.

...I want...my power.

I want it more than anything else. I want the life to move to the fun parts. I want the things that I do to have real effect upon reality. A person like me is not meant to be at the bottom forever.

And today, I will make a small step towards proving that.

9:40am. I can't tell at all how months down the road the interviewing will go. Ideally the people would get it, but that idiot schizo plus trying to explain the self improvement loop to one of my acquaintances killed me. Because of that I do not have much confidence in my ability to highlight the importance of my work. The guys at the ML sub seem to be perfectly content with PyTorch and Tensorflow.

I really did think that there would be at least a few other people enamored with the approach I am taking.

...It is a pity.

9:45am. Well, v0.2 needs to be given a shot. It is true that programming the way I did in 2018 in Spiral became unbearable even for me. If that is the case for me, then things have to be much worse for other people.

9:50am. Having confidence or not does not matter. I might not be good at talking to morons, but I have at least enough brains to recognize this fact. The only thing that matters is recognizing what I have to do and then doing it. I need to keep going and try. I need to risk failure.

One vision, one purpose.

You can recognize mediocrities by their parroting of the wisdom that technical skills do not matter. It would absolutely shock the normies if they were to ever find out that social skills were just a skill and that you can allocate autism points into it just as anything else."